+++ src/cascade/CMakeLists.txt
@@ -8,13 +8,13 @@ file(GLOB SOURCE_FILES
 	cascade.cc
 )
 
-file (GLOB_RECURSE SOURCE_DIRECTORY "../src/*.cc")
+file (GLOB_RECURSE SOURCE_DIRECTORY "../*.cc")
 
 add_library(libcascade SHARED STATIC ${SOURCE_FILES} ${SOURCE_DIRECTORY})
 set_target_properties(libcascade PROPERTIES PREFIX "")
 
 # build mongoose (TODO: remove from source)
-add_library(mongoose SHARED STATIC ../ext/mongoose/mongoose.c)
+add_library(mongoose SHARED STATIC ../../ext/mongoose/mongoose.c)
 
 target_link_libraries(libcascade verilog mongoose)
