+++ test/regression/remote.cc
@@ -28,8 +28,8 @@
 // OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+#include "cascade/cascade.h"
 #include "gtest/gtest.h"
-#include "cascade.h"
 #include "harness.h"
 
 using namespace cascade;
