+++ tools/cascade.cc
@@ -32,7 +32,7 @@
 #include <signal.h>
 #include "cl.h"
 #include "base/system/system.h"
-#include "cascade.h"
+#include "cascade/cascade.h"
 
 using namespace cl;
 using namespace cascade;
